Add inline mode for query diff
