+++ js/src/components/query/QueryDiffResultView.tsx
@@ -6,7 +6,7 @@ import {
   QueryDiffViewOptions,
   QueryPreviewChangeParams,
 } from "@/lib/api/adhocQuery";
-import { Box, Center, Flex, forwardRef } from "@chakra-ui/react";
+import { Center, Checkbox, Flex, forwardRef } from "@chakra-ui/react";
 import { useMemo } from "react";
 import { toDataDiffGrid } from "./querydiff";
 import { toValueDiffGrid as toQueryDiffJoinGrid } from "../valuediff/valuediff";
@@ -16,6 +16,8 @@ import { Run } from "@/lib/api/types";
 import { EmptyRowsRenderer, ScreenshotDataGrid } from "../data-grid/ScreenshotDataGrid";
 import { RunResultViewProps } from "../run/types";
 import { RunToolbar } from "../run/RunToolbar";
+import { DiffDislayModeSwitch } from "./ToggleSwitch";
+import { ChangedOnlyCheckbox } from "./ChangedOnlyCheckbox";
 
 export interface QueryDiffResultViewProps
   extends RunResultViewProps<QueryDiffParams, QueryDiffResult, QueryDiffViewOptions> {
@@ -38,6 +40,7 @@ const PrivateQueryDiffResultView = (
   const primaryKeys = useMemo(() => viewOptions?.primary_keys || [], [viewOptions]);
   const changedOnly = useMemo(() => viewOptions?.changed_only || false, [viewOptions]);
   const pinnedColumns = useMemo(() => viewOptions?.pinned_columns || [], [viewOptions]);
+  const displayMode = useMemo(() => viewOptions?.display_mode || "inline", [viewOptions]);
 
   const gridData = useMemo(() => {
     const handlePrimaryKeyChanged = (primaryKeys: string[]) => {
@@ -66,6 +69,7 @@ const PrivateQueryDiffResultView = (
       onPinnedColumnsChange: handlePinnedColumnsChanged,
       baseTitle,
       currentTitle,
+      displayMode,
     });
   }, [
     run,
@@ -73,6 +77,7 @@ const PrivateQueryDiffResultView = (
     changedOnly,
     primaryKeys,
     pinnedColumns,
+    displayMode,
     onViewOptionsChanged,
     baseTitle,
     currentTitle,
@@ -129,8 +134,29 @@ const PrivateQueryDiffResultView = (
         viewOptions={viewOptions}
         onAddToChecklist={onAddToChecklist}
         onViewOptionsChanged={onViewOptionsChanged}
-        warnings={warnings}
-      />
+        warnings={warnings}>
+        <DiffDislayModeSwitch
+          displayMode={displayMode}
+          onDisplayModeChanged={(displayMode) => {
+            if (onViewOptionsChanged) {
+              onViewOptionsChanged({
+                ...viewOptions,
+                display_mode: displayMode,
+              });
+            }
+          }}
+        />
+
+        <ChangedOnlyCheckbox
+          changedOnly={viewOptions?.changed_only}
+          onChange={() => {
+            const changedOnly = !viewOptions?.changed_only;
+            if (onViewOptionsChanged) {
+              onViewOptionsChanged({ ...viewOptions, changed_only: changedOnly });
+            }
+          }}
+        />
+      </RunToolbar>
       <ScreenshotDataGrid
         ref={ref}
         style={{ blockSize: "auto", maxHeight: "100%", overflow: "auto" }}
@@ -162,6 +188,7 @@ const PrivateQueryDiffJoinResultView = (
 ) => {
   const changedOnly = useMemo(() => viewOptions?.changed_only || false, [viewOptions]);
   const pinnedColumns = useMemo(() => viewOptions?.pinned_columns || [], [viewOptions]);
+  const displayMode = useMemo(() => viewOptions?.display_mode || "inline", [viewOptions]);
 
   const gridData = useMemo(() => {
     const handlePinnedColumnsChanged = (pinnedColumns: string[]) => {
@@ -209,7 +236,6 @@ const PrivateQueryDiffJoinResultView = (
         <RunToolbar
           run={run}
           viewOptions={viewOptions}
-          onAddToChecklist={onAddToChecklist}
           onViewOptionsChanged={onViewOptionsChanged}
           warnings={warnings}
         />
@@ -223,10 +249,30 @@ const PrivateQueryDiffJoinResultView = (
       <RunToolbar
         run={run}
         viewOptions={viewOptions}
-        onAddToChecklist={onAddToChecklist}
         onViewOptionsChanged={onViewOptionsChanged}
-        warnings={warnings}
-      />
+        warnings={warnings}>
+        <DiffDislayModeSwitch
+          displayMode={displayMode}
+          onDisplayModeChanged={(displayMode) => {
+            if (onViewOptionsChanged) {
+              onViewOptionsChanged({
+                ...viewOptions,
+                display_mode: displayMode,
+              });
+            }
+          }}
+        />
+
+        <ChangedOnlyCheckbox
+          changedOnly={viewOptions?.changed_only}
+          onChange={() => {
+            const changedOnly = !viewOptions?.changed_only;
+            if (onViewOptionsChanged) {
+              onViewOptionsChanged({ ...viewOptions, changed_only: changedOnly });
+            }
+          }}
+        />
+      </RunToolbar>
       <ScreenshotDataGrid
         ref={ref}
         style={{ blockSize: "auto", maxHeight: "100%", overflow: "auto" }}
