+++ js/src/components/run/RunToolbar.tsx
@@ -1,7 +1,6 @@
-import { AddIcon, WarningIcon } from "@chakra-ui/icons";
-import { Box, Button, Checkbox, Flex, IconButton, Spacer, Tooltip, VStack } from "@chakra-ui/react";
+import { WarningIcon } from "@chakra-ui/icons";
+import { Box, Flex, Spacer, VStack } from "@chakra-ui/react";
 import { RunResultViewProps } from "./types";
-import { Run } from "@/lib/api/types";
 
 interface DiffViewOptions {
   changed_only?: boolean;
@@ -9,23 +8,13 @@ interface DiffViewOptions {
 
 interface RunToolbarProps<PT, RT, VO> extends RunResultViewProps<PT, RT, VO> {
   warnings?: string[];
-  onAddToChecklist?: (run: Run<PT, RT>) => void;
+  children?: React.ReactNode;
 }
 
 export const RunToolbar = <PT, RT>({
-  run,
   warnings,
-  viewOptions,
-  onAddToChecklist,
-  onViewOptionsChanged,
+  children,
 }: RunToolbarProps<PT, RT, DiffViewOptions>) => {
-  const toggleChangedOnly = () => {
-    const changedOnly = !viewOptions?.changed_only;
-    if (onViewOptionsChanged) {
-      onViewOptionsChanged({ ...viewOptions, changed_only: changedOnly });
-    }
-  };
-
   return (
     <Flex
       borderBottom="1px solid lightgray"
@@ -42,20 +31,7 @@ export const RunToolbar = <PT, RT>({
         ))}
       </VStack>
       <Spacer minHeight="32px" />
-      <Checkbox isChecked={viewOptions?.changed_only} onChange={toggleChangedOnly}>
-        Changed only
-      </Checkbox>
-      {onAddToChecklist && (
-        <Button
-          marginBlock="5px"
-          size="sm"
-          colorScheme="blue"
-          onClick={() => {
-            onAddToChecklist(run);
-          }}>
-          Add to Checklist
-        </Button>
-      )}
+      {children}
     </Flex>
   );
 };
